Shuffle the arrays instead of sorting them with a function that returns a random value.
When using .sort(Math.random() - 0.5) the mountains tends to stay in place. The leftmost
mountain will stay in place 3/10 of the time, while the rightmost will stay in place
1/2 of the time. This must be due to the sorting algorithm used in the javascript
implementation.

Shuffling the arrays instead make sure that the placement of each mountain is
completely random. You could of course say that the mountains tendency to appear in
the same place is a feature, but it was not an intended feature.